requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
ultralytics==8.3.95
opencv-python-headless==4.8.0.76
Pillow==9.5.0
numpy==1.24.3
pydantic==2.1.1
tenacity==8.2.2
